refact: 모달과 컴포넌트 분리
#266
