+++ src/containers/team/teamMember/types.ts
@@ -1,0 +1,7 @@
+import { Member } from '@/types';
+
+export interface MemberModalProps {
+  member: Member;
+  isOpen: boolean;
+  onClose: () => void;
+}
